refactor: Modularize Dashboard component
- Extract SystemOverviewStats component
- Rename and enhance EndpointFilters component
- Improve component separation and readability
